// block-performance-1.0.0/src/library/web-client/get-file/type.ts
export interface Input {
    url: string;
    headers?: { name: string; value: string }[];
    save_to_file: string;
}

export interface OutputUnexpectedResponse {
    status_code: number;
    status_message: string;
    headers: { name: string; value: string }[];
}

export interface OutputOk {
    status_code: number;
    status_message: string;
    headers: { name: string; value: string }[];
}

export interface OutputFail {}

export interface Callback<R> {
    unexpected_response: (output: OutputUnexpectedResponse) => R;
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}
// ...
export type Output = OutputOk;